Refactoring typography for the legacy Website Next components
