+++ site/pages/event/style.css
@@ -44,7 +44,8 @@
 }
 
 .h1 {
-  composes: h1 from "../../components/typography/style.css";
+  composes: serif from "../../css/typography/_fonts.css";
+  composes: fontL from "../../css/typography/_fonts.css";
   text-align: center;
 }
 
@@ -83,7 +84,6 @@
 }
 
 .fullDetailsLink {
-  composes: aBold from "../../components/typography/style.css";;
   display: block;
   text-decoration: none;
   margin-top: 0.8em;
