+++ site/pages/events/style.css
@@ -1,7 +1,8 @@
 @value badgerWhite from "../../css/_colors.css";
 
 .h1 {
-  composes: h1 from "../../components/typography/style.css";
+  composes: serif from "../../css/typography/_fonts.css";
+  composes: fontL from "../../css/typography/_fonts.css";
   text-align: center;
   padding-top: 10px;
   background-color: badgerWhite;
