+++ site/pages/job/style.css
@@ -1,12 +1,19 @@
 @value tablet from "../../components/variables/breakpoints.css";
 @value badgerWhite, badgerBlack, badgerRed from "../../css/_colors.css";
 
+.jobTitle {
+  composes: serif from "../../css/typography/_fonts.css";
+  composes: fontM from "../../css/typography/_fonts.css";
+  text-align: left;
+}
+
 .background {
   background-color: badgerWhite;
 }
 
 .noteTitle {
-  composes: h2 from "../../components/typography/style.css";
+  composes: serif from "../../css/typography/_fonts.css";
+  composes: fontM from "../../css/typography/_fonts.css";
   font-size: 1.333em !important;
 }
